Improve Packaging (#37)
* Add `widgets` packages
* Organized `ui` package feature base
* Rename `core` package to `utils`
* Put student and exam entities in `entities` packages
