--- app/src/main/java/com/application/presence_absence/ui/core/CheckBoxItemView.kt
@@ -1,3 +1,0 @@
-package com.application.presence_absence.ui.core
-
-open class CheckBoxItemView(val text: String, var checked: Boolean)
