--- app/src/main/java/com/application/presence_absence/ui/examList/filter/ExamFilterView.kt
@@ -1,8 +1,0 @@
-package com.application.presence_absence.ui.examList.filter
-
-data class ExamFilterView(
-    val examPlace: List<String> = listOf(),
-    // TODO: update after implement
-    val examTime: String? = null,
-    val examState: List<String> = listOf()
-)
