--- app/src/main/java/com/application/presence_absence/ui/examList/filter/college/FakeCollegeList.kt
@@ -1,21 +1,0 @@
-package com.application.presence_absence.ui.examList.filter.college
-
-import com.application.presence_absence.ui.core.CheckBoxItemView
-
-object FakeCollegeList {
-    val list = listOf(
-        CheckBoxItemView(text = "مهندسی", checked = false),
-        CheckBoxItemView(text = "دانشکده ادبیات", checked = false),
-        CheckBoxItemView(text = "2دانشکده مهندسی", checked = false),
-        CheckBoxItemView(text = "دانشکده مهندسی3", checked = false),
-        CheckBoxItemView(text = "دانشکده مهندسی4", checked = false),
-        CheckBoxItemView(text = "دانشکده مهندسی5", checked = false),
-        CheckBoxItemView(text = "دانشکده مهندسی6", checked = false),
-        CheckBoxItemView(text = "دانشکده مهندسی7", checked = false),
-        CheckBoxItemView(text = "8دانشکده مهندسی", checked = false),
-        CheckBoxItemView(text = "9دانشکده مهندسی", checked = false),
-        CheckBoxItemView(text = "11دانشکده مهندسی", checked = false),
-        CheckBoxItemView(text = "دانشکده مهندسی12", checked = false),
-        CheckBoxItemView(text = "دانشکده مهندسی14", checked = false),
-    )
-}
