+++ app/src/main/java/com/application/presence_absence/ui/features/examList/entities/ExamFilterStateView.kt
@@ -1,0 +1,8 @@
+package com.application.presence_absence.ui.features.examList.entities
+
+data class ExamFilterStateView(
+    val examPlace: List<String> = listOf(),
+    // TODO: update after implement
+    val examTime: String? = null,
+    val examState: List<String> = listOf()
+)
