+++ app/src/main/java/com/application/presence_absence/ui/features/examList/entities/ExamView.kt
@@ -1,0 +1,15 @@
+package com.application.presence_absence.ui.features.examList.entities
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class ExamView(
+    val name: String,
+    val className: String,
+    val collegeName: String,
+    val day: String,
+    val hour: String,
+    val studentCount: String,
+    val state: ExamStatus
+) : Parcelable
