+++ app/src/main/java/com/application/presence_absence/ui/features/studentList/FakeStudentList.kt
@@ -1,0 +1,87 @@
+package com.application.presence_absence.ui.features.studentList
+
+import com.application.presence_absence.ui.features.studentList.entities.StudentAttendanceState
+import com.application.presence_absence.ui.features.studentList.entities.StudentView
+
+object FakeStudentList {
+    val list = listOf(
+        StudentView(
+            name = "ریاضی",
+            iconUrl = null,
+            idNumber = "9812762339",
+            attendance = StudentAttendanceState.NOT_SET
+        ),
+        StudentView(
+            name = "ریاضی",
+            iconUrl = null,
+            idNumber = "9812762339",
+            attendance = StudentAttendanceState.PRESENCE
+        ),
+        StudentView(
+            name = "ریاضی",
+            iconUrl = null,
+            idNumber = "9812762339",
+            attendance = StudentAttendanceState.ABSENCE
+        ),
+        StudentView(
+            name = "ریاضی",
+            iconUrl = null,
+            idNumber = "9812762339",
+            attendance = StudentAttendanceState.NOT_SET
+        ),
+        StudentView(
+            name = "ریاضی",
+            iconUrl = null,
+            idNumber = "9812762339",
+            attendance = StudentAttendanceState.NOT_SET
+        ),
+        StudentView(
+            name = "ریاضی",
+            iconUrl = null,
+            idNumber = "9812762339",
+            attendance = StudentAttendanceState.PRESENCE
+        ),
+        StudentView(
+            name = "ریاضی",
+            iconUrl = null,
+            idNumber = "9812762339",
+            attendance = StudentAttendanceState.ABSENCE
+        ),
+        StudentView(
+            name = "ریاضی",
+            iconUrl = null,
+            idNumber = "9812762339",
+            attendance = StudentAttendanceState.NOT_SET
+        ),
+        StudentView(
+            name = "ریاضی",
+            iconUrl = null,
+            idNumber = "9812762339",
+            attendance = StudentAttendanceState.ABSENCE
+        ),
+        StudentView(
+            name = "ریاضی",
+            iconUrl = null,
+            idNumber = "9812762339",
+            attendance = StudentAttendanceState.NOT_SET
+        ),
+        StudentView(
+            name = "ریاضی",
+            iconUrl = null,
+            idNumber = "9812762339",
+            attendance = StudentAttendanceState.PRESENCE
+        ),
+        StudentView(
+            name = "ریاضی",
+            iconUrl = null,
+            idNumber = "9812762339",
+            attendance = StudentAttendanceState.NOT_SET
+        ),
+        StudentView(
+            name = "ریاضی",
+            iconUrl = null,
+            idNumber = "9812762339",
+            attendance = StudentAttendanceState.ABSENCE
+        ),
+    )
+}
