+++ app/src/main/java/com/application/presence_absence/ui/features/studentList/entities/StudentFilterStateView.kt
@@ -1,0 +1,6 @@
+package com.application.presence_absence.ui.features.studentList.entities
+
+data class StudentFilterStateView(
+    val showPresents: Boolean? = null,
+    val showAbsents: Boolean? = null
+)
