+++ app/src/main/java/com/application/presence_absence/ui/features/studentList/entities/StudentView.kt
@@ -1,0 +1,8 @@
+package com.application.presence_absence.ui.features.studentList.entities
+
+data class StudentView(
+    val name: String,
+    val iconUrl: String?,
+    val idNumber: String,
+    var attendance: StudentAttendanceState
+)
