--- app/src/main/java/com/application/presence_absence/ui/studentList/FakeStudentList.kt
@@ -1,19 +1,0 @@
-package com.application.presence_absence.ui.studentList
-
-object FakeStudentList {
-    val list = listOf(
-        StudentView(name = "ریاضی", iconUrl = null, idNumber = "9812762339", attendance = StudentAttendance.NOT_SET),
-        StudentView(name = "ریاضی", iconUrl = null, idNumber = "9812762339", attendance = StudentAttendance.PRESENCE),
-        StudentView(name = "ریاضی", iconUrl = null, idNumber = "9812762339", attendance = StudentAttendance.ABSENCE),
-        StudentView(name = "ریاضی", iconUrl = null, idNumber = "9812762339", attendance = StudentAttendance.NOT_SET),
-        StudentView(name = "ریاضی", iconUrl = null, idNumber = "9812762339", attendance = StudentAttendance.NOT_SET),
-        StudentView(name = "ریاضی", iconUrl = null, idNumber = "9812762339", attendance = StudentAttendance.PRESENCE),
-        StudentView(name = "ریاضی", iconUrl = null, idNumber = "9812762339", attendance = StudentAttendance.ABSENCE),
-        StudentView(name = "ریاضی", iconUrl = null, idNumber = "9812762339", attendance = StudentAttendance.NOT_SET),
-        StudentView(name = "ریاضی", iconUrl = null, idNumber = "9812762339", attendance = StudentAttendance.ABSENCE),
-        StudentView(name = "ریاضی", iconUrl = null, idNumber = "9812762339", attendance = StudentAttendance.NOT_SET),
-        StudentView(name = "ریاضی", iconUrl = null, idNumber = "9812762339", attendance = StudentAttendance.PRESENCE),
-        StudentView(name = "ریاضی", iconUrl = null, idNumber = "9812762339", attendance = StudentAttendance.NOT_SET),
-        StudentView(name = "ریاضی", iconUrl = null, idNumber = "9812762339", attendance = StudentAttendance.ABSENCE),
-    )
-}
