--- app/src/main/java/com/application/presence_absence/ui/studentList/StudentAttendance.kt
@@ -1,10 +1,0 @@
-package com.application.presence_absence.ui.studentList
-
-import androidx.annotation.StringRes
-import com.application.presence_absence.R
-
-enum class StudentAttendance(@StringRes val titleId: Int) {
-    PRESENCE(R.string.label_present),
-    ABSENCE(R.string.label_absent),
-    NOT_SET(-1)
-}
