--- app/src/main/java/com/application/presence_absence/ui/studentList/StudentFilterView.kt
@@ -1,6 +1,0 @@
-package com.application.presence_absence.ui.studentList
-
-data class StudentFilterView(
-    val showPresents: Boolean? = null,
-    val showAbsents: Boolean? = null
-)
