--- app/src/main/java/com/application/presence_absence/ui/studentList/StudentView.kt
@@ -1,8 +1,0 @@
-package com.application.presence_absence.ui.studentList
-
-data class StudentView(
-    val name: String,
-    val iconUrl: String?,
-    val idNumber: String,
-    var attendance: StudentAttendance
-)
