+++ app/src/main/java/com/application/presence_absence/ui/widgets/CheckBoxItemView.kt
@@ -1,0 +1,3 @@
+package com.application.presence_absence.ui.widgets
+
+open class CheckBoxItemView(val text: String, var checked: Boolean)
