+++ app/src/main/res/layout/item_exam.xml
@@ -8,11 +8,11 @@
 
         <variable
             name="exam"
-            type="com.application.presence_absence.ui.examList.ExamView" />
+            type="com.application.presence_absence.ui.features.examList.entities.ExamView" />
 
         <variable
             name="tag"
-            type="com.application.presence_absence.ui.examList.ExamStatusTagView" />
+            type="com.application.presence_absence.ui.features.examList.entities.ExamStatusTagView" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
